+++ sig/datadog/error_tracking/configuration.rbs
@@ -1,6 +1,10 @@
 module Datadog
   module ErrorTracking
     module Configuration
+      module Settings
+        def self.extended: (untyped base) -> untyped
+        def self.add_settings!: (untyped base) -> untyped
+      end
     end
   end
 end
