--- sig/datadog/error_tracking/configuration/settings.rbs
@@ -1,10 +1,0 @@
-module Datadog
-  module ErrorTracking
-    module Configuration
-      module Settings
-        def self.extended: (untyped base) -> untyped
-        def self.add_settings!: (untyped base) -> untyped
-      end
-    end
-  end
-end
